+++ lib/Report/Generator/EnvGenerator.php
@@ -13,9 +13,15 @@
 namespace PhpBench\Report\Generator;
 
 use PhpBench\Dom\Document;
+use PhpBench\Expression\Ast\BooleanNode;
+use PhpBench\Expression\Ast\PhpValueFactory;
+use PhpBench\Expression\Ast\StringNode;
 use PhpBench\Model\SuiteCollection;
 use PhpBench\Registry\Config;
 use PhpBench\Report\GeneratorInterface;
+use PhpBench\Report\Model\Report;
+use PhpBench\Report\Model\Reports;
+use PhpBench\Report\Model\Table;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
 /**
@@ -40,55 +46,32 @@ class EnvGenerator implements GeneratorInterface
     /**
      * {@inheritdoc}
      */
-    public function generate(SuiteCollection $suiteCollection, Config $config): Document
+    public function generate(SuiteCollection $suiteCollection, Config $config): Reports
     {
-        $document = new Document();
-        $reportsEl = $document->createRoot('reports');
-        $reportsEl->setAttribute('name', 'table');
-        $reportEl = $reportsEl->appendElement('report');
+        $title = isset($config['title']) ? $config['title'] : null;
+        $description = isset($config['description']) ? $config['description'] : null;
 
-        if (isset($config['title'])) {
-            $reportEl->setAttribute('title', $config['title']);
-        }
-
-        if (isset($config['description'])) {
-            $reportEl->appendElement('description', $config['description']);
-        }
+        $tables = [];
 
         foreach ($suiteCollection as $suite) {
-            $tableEl = $reportEl->appendElement('table');
-            $colsEl = $tableEl->appendElement('cols');
 
-            foreach (['provider', 'key', 'value'] as $colName) {
-                $col = $colsEl->appendElement('col');
-                $col->setAttribute('name', $colName);
-                $col->setAttribute('label', $colName);
-            }
-
-            $tableEl->setAttribute('title', sprintf(
+            $title = sprintf(
                 'Suite #%s %s', $suite->getUuid(), $suite->getDate()->format('Y-m-d H:i:s')
-            ));
+            );
 
-            $groupEl = $tableEl->appendElement('group');
-            $groupEl->setAttribute('name', 'body');
-
+            $rows = [];
             foreach ($suite->getEnvInformations() as $envInformation) {
                 foreach ($envInformation as $key => $value) {
-                    $rowEl = $groupEl->appendElement('row');
-
-                    $cellEl = $rowEl->appendElement('cell', $envInformation->getName());
-                    $valueEl = $cellEl->appendElement('value', $envInformation->getName());
-                    $cellEl->setAttribute('name', 'provider');
-                    $cellEl = $rowEl->appendElement('cell');
-                    $valueEl = $cellEl->appendElement('value', $key);
-                    $cellEl->setAttribute('name', 'key');
-                    $cellEl = $rowEl->appendElement('cell');
-                    $valueEl = $cellEl->appendElement('value', is_bool($value) ? $value ? 'yes' : 'no' : $value);
-                    $cellEl->setAttribute('name', 'value');
+                    $rows[] = [
+                        'provider' => new StringNode($envInformation->getName()),
+                        'key' => new StringNode($key),
+                        'value' => PhpValueFactory::fromNumber($value)
+                    ];
                 }
             }
+            $tables[] = Table::fromRowArray($rows, $title);
         }
 
-        return $document;
+        return Reports::fromReport(new Report($tables, $title, $description));
     }
 }
